Manual control improvements

+++ sources/domain/manual_controller/joystick_controller.cpp
@@ -30,7 +30,7 @@ JoystickController::JoystickController(QObject* parent):
 #endif
 }
 
-float JoystickController::value(int axis) const
+double JoystickController::value(int axis) const
 {
     switch (axis)
     {
