+++ sources/domain/manual_controller/joystick_controller.h
@@ -14,14 +14,14 @@ namespace domain
     public:
         explicit JoystickController(QObject* parent = nullptr);
 
-        float value(int axis) const;
+        double value(int axis) const;
         int deviceId() const;
 
     public slots:
         void setDeviceId(int deviceId);
 
     signals:
-        void valueChanged(int axis, float value);
+        void valueChanged(int axis, double value);
 
     private:
         QGamepad* m_gamepad;
